backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
  "anywidget>=0.9.18",
  "bqplot>=0.12.45",
  "branca>=0.8.1",
  "earthengine-api>=1.6.3",
  "ee>=0.2",
  "eerepr>=0.1.2",
  "flask[async]>=3.1.1",
  "flask-cors>=6.0.1",
  "folium>=0.20.0",
  "geemap>=0.36.1",
  "geocoder>=1.38.1",
  "geopy>=2.4.1",
  "jupyter>=1.1.1",
  "jupyterlab-vim>=4.1.4",
  "matplotlib>=3.10.5",
  "numpy>=2.3.2",
  "openai>=1.99.9",
  "pandas>=2.3.1",
  "plotly>=6.3.0",
  "pydantic-ai>=0.7.2",
  "python-dotenv>=1.0.1",
  "requests>=2.32.4",
  "traitlets>=5.14.3",
  "xyzservices>=2025.4.0",
]


